news filtering by current region, include subregions, if result is less than 3 news, it will include parent region until grater than 3, news from same region are sorted by date
