+++ assets/js/components/ProjectsNews.vue
@@ -1,0 +1,13 @@
+<script>
+
+import ProjectNews from './BaseNews';
+import ProjectsMixin from './mixins/Projects';
+
+export default ProjectNews.extend({
+
+  mixins: [
+    ProjectsMixin,
+  ],
+
+});
+</script>
