+++ assets/js/root-instances.js
@@ -47,7 +47,7 @@ export const Partners = Base.extend({
     xmap: components.PartnersMap,
     donors: components.PartnersDonors,
     beneficiaries: components.PartnersBeneficiaries,
-    news: components.PartnerNews,
+    news: components.PartnersNews,
     programme_partners: components.PartnersDonorProgrammes,
     project_partners: components.PartnersDonorProjects,
     xsummary: components.PartnersSummary,
@@ -65,7 +65,7 @@ export const Projects = Base.extend({
     xmap: components.ProjectsMap,
     beneficiaries: components.ProjectsBeneficiaries,
     xsummary: components.ProjectsSummary,
-    news: components.News,
+    news: components.ProjectsNews,
     projects: components.Projects,
   },
 });
